+++ app/models/project_favorite.rb
@@ -1,2 +1,7 @@
 class ProjectFavorite < ActiveRecord::Base
+
+  # Model Relationships
+  belongs_to :project
+  belongs_to :user
+
 end