+++ db/migrate/20150414145657_add_position_to_project_favorites.rb
@@ -1,0 +1,5 @@
+class AddPositionToProjectFavorites < ActiveRecord::Migration
+  def change
+    add_column :project_favorites, :position, :integer, null: false, default: 0
+  end
+end